SIG-304 Se agrega validacion al momento de eliminar diagnostico que no posea pcl

+++ src/app/modulos/scm/components/formulario-scm/formulario-scm.component.ts
@@ -62,6 +62,9 @@ export interface TreeNode {
 })
 export class FormularioScmComponent implements OnInit, OnDestroy {
 
+    listaPCL;
+    itemInPCL:boolean=false;
+
     empresasList: Empresa[];
     styleMap: { [key: string]: string } = {};
     value;
@@ -1033,26 +1036,35 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
     }
 
     async deleteDiagnostico(id) {
+        
+        console.log("diagnostico:",id)
         this.msgs = [];
         try {
-            if (await this.confirmService.confirmDiagnostico()){
-                let resp = await this.scmService.deleteDiagnosticos(id);
-                if (resp) {
-                    this.msgs.push({
-                        severity: "error",
-                        summary: "Diagnostico",
-                        detail: `Su Diagnostico fue eliminado`,
-                    });
-                    this.onCloseModalDianostico();
+            await this.validarPCL()
+            if(!this.itemInPCL){               
+            
+                if (await this.confirmService.confirmDiagnostico()){
+                    let resp = await this.scmService.deleteDiagnosticos(id);
+                    if (resp) {
+                        this.msgs.push({
+                            severity: "error",
+                            summary: "Diagnostico",
+                            detail: `Su Diagnostico fue eliminado`,
+                        });
+                        this.onCloseModalDianostico();
+                    }
+                }            
+                else {
+                    this.msgs = [
+                        { severity: "info", summary: "Cancelado", detail: "usted cancelo la eliminación" }
+                    ];
                 }
-            }            
-            else {
+            }
+            else{
                 this.msgs = [
-                    { severity: "info", summary: "Cancelado", detail: "usted cancelo la eliminación" }
+                    { severity: "info", summary: "Cancelado", detail: "El diagnostico contiene PCL activos" }
                 ];
             }
-
-           
 
         } catch (error) {
             console.log(error)
@@ -1090,7 +1102,7 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
 
     async deleteSeguimiento(id) {
         this.msgs = [];
-        try {
+        try {           
             if (await this.confirmService.confirmSeguimiento()){
                 let resp = await this.scmService.deleteSeguimiento(id);
                 if (resp) {
@@ -1163,4 +1175,26 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
         this.modalRecomendatios = true;
     }
 
+    
+    ok(event){
+        console.log("------------------------------------------------",event)
+        console.log("************************************************",this.diagnosticoList)
+        console.log("////////////////////////////////////////////////",this.diagSelect)
+
+
+        this.listaPCL=event
+    }
+    validarPCL(){
+        this.itemInPCL=false;
+
+        // console.log(this.diagSelect.diagnostico)
+        this.listaPCL.forEach(item => {
+           if(item.diagnostic.label == this.diagSelect.diagnostico){
+            this.itemInPCL=true;
+            // return true;
+           }
+       }); 
+    //    console.log(this.itemInPCL)
+    }
+    
 }
